+++ src/node/worker/worker.ts
@@ -1,24 +1,14 @@
 import {
-  getCronTaskJob,
-  getJobCollection,
-  getSimpleJob,
-  updateJob,
-} from "../data/actions.ts";
-import {
-  IJob,
-  IJobsCronTask,
-  IJobsSimple,
-  isJobCronTask,
-} from "../../common/model.ts";
-import { CronDef, ILogger, JobDef, getLogger, getOptions } from "../setup.ts";
-import {
   captureException,
-  runWithAsyncContext,
   getCurrentHub,
-  captureCheckIn,
+  runWithAsyncContext,
 } from "@sentry/node";
 import { formatDuration, intervalToDuration } from "date-fns";
+import { IJobsCronTask, IJobsSimple } from "../../common/model.ts";
+import { getCronTaskJob, getSimpleJob, updateJob } from "../data/actions.ts";
+import { CronDef, ILogger, JobDef, getLogger, getOptions } from "../setup.ts";
 import { workerId } from "./heartbeat.ts";
+import { notifySentryJobEnd, notifySentryJobStart } from "./sentry.ts";
 
 function getJobSimpleDef(job: IJobsSimple): JobDef | null {
   const options = getOptions();
@@ -252,80 +242,3 @@ export async function reportJobCrash(
     );
   }
 }
-
-const notifySentryJobStart = async (job: IJob) => {
-  if (!isJobCronTask(job)) {
-    return;
-  }
-  const monitorConfig = getSentryMonitorConfig(job.name);
-  if (!monitorConfig) {
-    getLogger().error(
-      { _id: job._id },
-      `unexpected: could not find cron definition`,
-    );
-    return;
-  }
-  const checkInId = captureCheckIn(
-    {
-      monitorSlug: job.name,
-      status: "in_progress",
-    },
-    monitorConfig,
-  );
-  await getJobCollection().findOneAndUpdate(
-    { _id: job._id },
-    { $set: { sentry_id: checkInId } },
-  );
-};
-
-const notifySentryJobEnd = async (job: IJob, isSuccess: boolean) => {
-  if (!isJobCronTask(job)) {
-    return;
-  }
-  const monitorConfig = getSentryMonitorConfig(job.name);
-  if (!monitorConfig) {
-    getLogger().error(
-      { _id: job._id },
-      `unexpected: could not find cron definition`,
-    );
-    return;
-  }
-  const dbJob = await getJobCollection().findOne({ _id: job._id });
-  if (!dbJob) {
-    getLogger().error({ _id: job._id }, `unexpected: could not find job`);
-    return;
-  }
-  if (!isJobCronTask(dbJob)) {
-    getLogger().error({ _id: job._id }, `unexpected: not a cron task`);
-    return;
-  }
-  const { sentry_id } = dbJob;
-  if (!sentry_id) {
-    getLogger().error({ _id: job._id }, `unexpected: no sentry_id`);
-    return;
-  }
-  captureCheckIn(
-    {
-      checkInId: sentry_id,
-      monitorSlug: job.name,
-      status: isSuccess ? "ok" : "error",
-    },
-    monitorConfig,
-  );
-};
-
-const getSentryMonitorConfig = (jobName: string) => {
-  const cronDefOpt = getOptions().crons[jobName];
-  if (!cronDefOpt) {
-    return null;
-  }
-  return {
-    schedule: {
-      type: "crontab",
-      value: cronDefOpt.cron_string,
-    },
-    checkinMargin: 5, // In minutes. Optional.
-    maxRuntime: cronDefOpt.maxRuntimeInMinutes ?? 60, // In minutes. Optional.
-    timezone: "Europe/Paris", // Optional.
-  } as const;
-};
